+++ tests/test_dynamic_graph_cpo_dod.cpp
@@ -872,7 +872,7 @@ TEST_CASE("dod CPO degree(g, u)", "[dynamic_graph][dod][cpo][degree]") {
         size_t idx = 0;
         
         for (auto u : vertices(g)) {
-            REQUIRE(static_cast<size_t>(degree(g, u)) == expected_degrees[idx]);
+            REQUIRE(degree(g, u) == expected_degrees[idx]);
             ++idx;
         }
     }
